request: remove RequestClient.Validate
After the last changes this would not affect performance. Not that it
matters too much anyway, while parsing the Result is kind slow, it is
not by far the impactful thing for performance, unless you are doing
maybe thousands of commands, but I think the IPC interface will be a
bottleneck much earlier than parsing the results.