request: do not append '/j' in RawRequest()
This change moves the prefix of '/j' of the `RawRequest()` to
`prepareRequests()` function instead. The prefix is used in Hyprland to
indicate that the request needs to return JSON format, and while we
always wants JSON format as response, someone using `RawRequest()`
directly may not want to do so.

While this is a breaking change for users of `RawRequest()`, the impact
should be minimal and will allow for more flexibility in future.

This also fixes a bug in batch mode, where we need to append '/j' in
every command, not just the first one. It didn't affect the e.g.
`Dispatch()` since it doesn't have a JSON output anyway, but it could
affect something in the future.