request: make 'j/' prefix for requests optional
Not every command supports JSON response (e.g.: Dispatch), and forcing
it in every command not only send an unnecessary request to Hyprland
IPC, but also can make it more fragile since if a future Hyprland
version implement JSON response for an existing command, this command
will break.

This commit adds a explicit `jsonResp` parameter to `prepareRequests()`
so we can differentiate between commands that return JSON response for
those that doesn't, and we will only prefix the request with 'j/' if
`jsonResp == true`.